online_course/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
httplib2==0.14.0
Jinja2==3.0.3
PyJWT==1.7.1
PyYAML==5.3.1
requests==2.22.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.16.0
sqlparse==0.4.2
urllib3==1.25.8
psycopg2-binary
celery==5.2.3
django-celery-results==2.3.0
python-decouple==3.6
flower==1.0.0
redis==4.1.4
pytest==7.0.1
pytest-django==4.5.2
